Move tokenize logic to data_utils
